Refactor: Extract type parser
