+++ tests/test_doxygenparser__objc.py
@@ -17,12 +17,6 @@ Tests for parsing Objective C from Doxygen XML files.
 
 import pytest
 
-import xml.etree.ElementTree as ET
-
-from asciidoxy.doxygenparser.objc import ObjectiveCParser
-from asciidoxy.doxygenparser import Driver as ParserDriver
-from .shared import assert_equal_or_none_if_empty
-
 
 @pytest.mark.parametrize("api_reference_set", [["objc/default"]])
 def test_parse_objc_interface(api_reference):
@@ -151,63 +145,3 @@ def test_parse_objc_block(api_reference):
 
     nested = param.type.nested[0]
     assert nested.name == "ADTrafficEvent"
-
-
-@pytest.fixture(params=[
-    "",
-    "nullable ",
-    "const ",
-    "__weak ",
-    "__strong ",
-    "nullable __weak ",
-    "nullable __strong ",
-])
-def objc_type_prefix(request):
-    return request.param
-
-
-@pytest.fixture(params=["", " *", " **", " * *"])
-def objc_type_suffix(request):
-    return request.param
-
-
-def test_parse_objc_type_from_text_simple(objc_type_prefix, objc_type_suffix):
-    parser = ParserDriver()
-    objc = ObjectiveCParser(parser)
-
-    type_element = ET.Element("type")
-    type_element.text = f"{objc_type_prefix}NSInteger{objc_type_suffix}"
-
-    type_ref = objc.parse_type(type_element)
-    assert type_ref is not None
-    assert type_ref.id is None
-    assert type_ref.kind is None
-    assert type_ref.language == "objc"
-    assert type_ref.name == "NSInteger"
-    assert_equal_or_none_if_empty(type_ref.prefix, objc_type_prefix)
-    assert_equal_or_none_if_empty(type_ref.suffix, objc_type_suffix)
-    assert len(type_ref.nested) == 0
-
-
-@pytest.mark.parametrize("type_with_space", [
-    "short int", "signed short", "signed short int", "unsigned short", "unsigned short int",
-    "signed int", "signed", "unsigned", "unsigned int", "long int", "signed long",
-    "signed long int", "unsigned long", "unsigned long int", "long long", "long long int",
-    "signed long long", "signed long long int", "unsigned long long", "unsigned long long int",
-    "signed char", "long double"
-])
-def test_parse_objc_type_with_space(type_with_space):
-    parser = ParserDriver()
-    objc = ObjectiveCParser(parser)
-
-    type_element = ET.Element("type")
-    type_element.text = type_with_space
-
-    type_ref = objc.parse_type(type_element)
-    assert type_ref is not None
-    assert not type_ref.id
-    assert not type_ref.kind
-    assert type_ref.language == "objc"
-    assert type_ref.name == type_with_space
-    assert type_ref.prefix is None
-    assert type_ref.suffix is None
